mswstr10: Add DLL and tests
Signed-off-by: Alex Henrie <alexhenrie24@gmail.com>
---
Required by Microsoft Access 2000.

mswstr10.dll was included in Windows XP, although it has been removed
from later Windows versions.

According to [1] this library contains Unicode collation tables.

Pages 395 and 396 of Internationalization with Visual Basic by Michael
Kaplan explain that the function with ordinal 2 is a copy of
LCMapStringW from kernel32. I guessed that one of the other functions
was a copy of CompareStringW and found that ordinal 1 is that function
through trial and error.

Ordinal 3 was more difficult to figure out because it does not appear to
correspond to any function in the rest of the Windows API. I tried
various numbers of arguments to see what would run without crashing,
then noticed that one of the arguments was a callback function, and
finally realized that it was enumerating the IDs and names of the
supported collation orders.

Unfortunately, even with this patch Microsoft Access 2000 still crashes
when creating a new database on the unimplemented function
msjet40.dll.154. That DLL was also distributed with Windows XP but there
is no documentation at all for it, so making it work is going to be
quite a bit harder. Still, having mswstr10.dll working means that Wine
users will only have to copy one closed-source DLL from Windows instead
of two in order to use Access 2000.

[1] http://archives.miloush.net/michkap/archive/2005/09/13/463962.html