winemenubuilder: Give Wine programs low file association priority
When multiple programs can open files of a particular type and the user
has not set a default program, the InitialPreference property helps
determine which program should open the file. Its default value is 1 and
a larger number indicates higher priority. Currently only KDE supports
InitialPreference but other desktop environments may add support in the
future.

In general, if both a native program and a Windows program can open a
file, the native program is the better choice. This is especially true
of plain text files, HTML files, image files, and other file types that
Wine's built-in programs can open. In the case where the user really
does want to use a Windows program by default, it's reasonable to
expect that they will set the default themselves.

Signed-off-by: Alex Henrie <alexhenrie24@gmail.com>